fix: extend SSR prefetch to blog list, blog detail, recruit detail
Same client waterfall issue identified in /news exists across:
- /blog list page
- /blog/[idx] detail page
- /recruit/[idx] detail page

Detail pages had a separate problem: generateMetadata fetches the
record server-side and the client refetches the same record after
hydration — every detail view was paying for two round trips.

Changes:
- /blog: page.tsx → server prefetch + HydrationBoundary, client
  logic moved to page-client.tsx (mirrors /news structure)
- /blog/[idx]: prefetchQuery in page.tsx, wrap client. getBlogDetailApi
  wrapped in react cache() so generateMetadata + prefetch share fetch.
- /recruit/[idx]: same pattern. getRecruitDetailApi cached.
- features/blog/query, features/recruit/query: drop "use client"
  directive so queryOptions factory is server-importable.
- Next 16 PageProps: params/searchParams now Promise — awaited.
